redo last commit

+++ dimabe_manufacturing/models/mrp_production.py
@@ -291,16 +291,19 @@ class MrpProduction(models.Model):
     @api.multi
     def fix_reserved(self):
         for item in self:
-            item.move_raw_ids.update({
-                'is_done': False,
-                'state': 'assigned'
-            })
+            if item.current_user_email not in ('desarrolloti@dimabe.cl','logistica@lainvernada.cl'):
+                raise models.UserError('Opcion no disponible')
+            else:
+                item.move_raw_ids.update({
+                    'is_done': False,
+                    'state': 'assigned'
+                })
 
-            for move in item.move_raw_ids:
-                if move.reserved_availability > 0:
-                    query = 'DELETE FROM stock_move_line where move_id = {}'.format(move.id)
-                    cr = self._cr
-                    cr.execute(query)
+                for move in item.move_raw_ids:
+                    if move.reserved_availability > 0:
+                        query = 'DELETE FROM stock_move_line where move_id = {}'.format(move.id)
+                        cr = self._cr
+                        cr.execute(query)
 
     # @api.multi
     # def action_cancel(self):
